encode tickets with Storable -> MIME::Base64
Ticket format has changed. Ticket strings are now the hash, then '--',
then the data. The data is the result of Storable::nfreeze($ticket) sent
through MIME::Base64::encode().  This makes it possible for subclasses
to add fields to the ticket, and all fields will be included in the
ticket checksum/hash